implemented external jumper for addressing
If the external jumper is closed the address is forced to 0x19. If the jumper is open then the eeprom will be read and if blank, the default address 0x18 will be used.

if the jumper is closed and the address is changed it will be stored in eeprom but the device will not use that address.